Added some more error checking to BaseDownloadManager

+++ LibPackageManager/Download/BaseDownloadManager.cs
@@ -3,12 +3,17 @@ using LibPackageManager.Provider;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace LibPackageManager.Download
 {
+    /// <summary>
+    /// Base for a download manager class for a given IRepositoryItem.
+    /// </summary>
+    /// <typeparam name="T">The IRepositoryItem that this class supports downloading.</typeparam>
     public abstract class BaseDownloadManager<T>
         where T : class, IRepositoryItem
     {
@@ -42,48 +47,73 @@ namespace LibPackageManager.Download
         /// Downloads and installs an item and its dependencies.
         /// </summary>
         /// <param name="item">The item to download.</param>
-        public async Task<string> DownloadItemAsync(T item)
+        /// <returns>True if the operation succeeded, false otherwise.</returns>
+        public async Task<bool> GetItemAsync(T item)
         {
-            List<Task> itemDownloadTasks = new List<Task>
+            // Check item is not null
+            if (item is null) throw new ArgumentNullException(nameof(item));
+
+            // Create download task list and start downloading item
+            List<Task<bool>> itemDownloadTasks = new()
             {
-                DownloadSingleItemAsync(item)
+                GetSingleItemAsync(item)
             };
 
+            // Check if this item has dependencies
             if (item is IDependentRepositoryItem depItem)
             {
+                // Start each dependency's download and add it to the download list
                 foreach (string key in depItem.Dependencies.Keys)
                 {
-                    itemDownloadTasks.Add(DownloadItemAsync(depItem.Dependencies[key] as T));
+                    itemDownloadTasks.Add(GetItemAsync(depItem.Dependencies[key] as T));
                 }
             }
 
+            // Wait until all tasks are completed
             await Task.WhenAll(itemDownloadTasks);
 
-            return null;
+            return !itemDownloadTasks.Select(x => x.Result).Contains(false);
         }
 
         /// <summary>
         /// Downloads and installs an item.
         /// </summary>
         /// <param name="item">The item to download and install.</param>
-        protected async Task DownloadSingleItemAsync(T item)
+        /// <exception cref="ArgumentNullException"
+        protected async Task<bool> GetSingleItemAsync(T item)
         {
-            if (item.IsDownloaded) return;
-            else if (item.DownloadUrl == null) return;
-            else if (DownloadsInProgress.Contains(item)) return;
+            if (item is null) throw new ArgumentNullException(nameof(item));
+
+            if (item.IsDownloaded) return true;
+            else if (item.DownloadUrl == null) return false;
+            else if (DownloadsInProgress.Contains(item)) return true;
 
             DownloadsInProgress.Add(item);
 
             string downloadPath = $"{downloadDir}/{Path.GetFileName(item.DownloadUrl)}";
 
             // Ensure directories exist by creating them
-            Directory.CreateDirectory(downloadDir);
-            Directory.CreateDirectory(installDir);
+            try
+            {
+                Directory.CreateDirectory(downloadDir);
+                Directory.CreateDirectory(installDir);
+            }
+            catch (Exception)
+            {
+                return false;
+            }
 
             // Download file
             using (WebClient client = new WebClient())
             {
-                await client.DownloadFileTaskAsync(item.DownloadUrl, downloadPath);
+                try
+                {
+                    await client.DownloadFileTaskAsync(item.DownloadUrl, downloadPath);
+                }
+                catch (Exception)
+                {
+                    return false;
+                }
             }
 
             // Install item from file
@@ -93,10 +123,21 @@ namespace LibPackageManager.Download
             item.InstallPath = $"{installDir}/{item.Id}";
 
             DownloadsInProgress.Remove(item);
+
+            // Remove downloaded file now we've installed its content
+            File.Delete(downloadPath);
+
+            return true;
         }
 
+        /// <summary>
+        /// Uninstalls an item.
+        /// </summary>
+        /// <param name="item">The item to uninstall.</param>
         public async Task RemoveItemAsync(T item)
         {
+            if (item is null) throw new ArgumentNullException(nameof(item));
+
             if (!item.IsDownloaded) return;
             if (!Directory.Exists($"{installDir}/{item.Id}")) return;
             await Task.Run(() => Directory.Delete($"{installDir}/{item.Id}", true));
